backend/app/services/ai_service.py
import os

from openai import OpenAI
from dotenv import load_dotenv


load_dotenv()

client = OpenAI(
    api_key=os.getenv("OPENAI_API_KEY")
)

def generate_ai_analysis(job_description: str):

    prompt = f"""
    Analyze this job description.

    Return:
    - summary
    - important skills
    - experience level

    Job Description:
    {job_description}
    """

    response = client.chat.completions.create(
        model="gpt-4.1-mini",
        messages=[
            {
                "role": "user",
                "content": prompt
            }
        ]
    )

    return response.choices[0].message.content